Mitigate reference error in dotnet 2.2 (#332)
* Mitigate reference error in dotnet 2.2

* Add comments on mitigation

* Fix run from package test case

+++ test/kudulite/containers.ts
@@ -263,13 +263,22 @@ export class KuduContainer {
         console.log(chalk.red.bold(`Failed to kill runtime container ${destContainerName}`));
       }
 
-      // Clean up docker image
-      console.log(chalk.yellow(`Cleaning up image ${baseImage}...`));
-      const rmiCommand = `docker rmi -f ${baseImage}`;
-      const rmiResult = shell.exec(rmiCommand);
-      if (rmiResult.code !== 0) {
-        console.log(chalk.red.bold(`Failed to remove runtime image ${baseImage}`));
+      // Clean up docker image ends with -python* -node* -java*
+      const isGenericImage = (
+        baseImage.endsWith(this.config.v2RuntimeVersion) || baseImage.endsWith(this.config.v3RuntimeVersion)
+      );
+      if (!isGenericImage) {
+        console.log(chalk.yellow(`Cleaning up image ${baseImage}...`));
+        const rmiCommand = `docker rmi -f ${baseImage}`;
+        const rmiResult = shell.exec(rmiCommand);
+        if (rmiResult.code !== 0) {
+          console.log(chalk.red.bold(`Failed to remove runtime image ${baseImage}`));
+        }
+      } else {
+        console.log(chalk.yellow(`Retained generatic image ${baseImage}.`));
       }
+
+      // Remove ports registry
       delete KuduContainer.ports[destContainerName];
     }
